+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Some code formats where not being properly colored.
 - Update all icons (Octicons v3.2.0).
 - Don't show references to nonexistent commits in issues.
+- If an issue is closed from a commit, show the commit in the event.
 
 ## [ForkHub v1.1.0] - 2015-11-06